feat(DTO): Creation of DTO for User table - PK issue
Creates a DTO that allowed me to customise *what* the API's GET call actually shows. Useful for security, BUT seems to be broken in regards to the primary key. Fix is likely a line or two, but it's late right now and I don't want to deal with it today.